src/requirements.txt
Flask==1.1.2
# gunicorn
requests
colorama
google-cloud-tasks
google-cloud-logging